Componentize Voter Registration a bit more.
Also fixes bug with some phones where image was not
stretching to the full width and looking funky
