+++ app/src/main/java/com/fueledbycaffeine/mivote/MainActivity.kt
@@ -1,24 +1,23 @@
 package com.fueledbycaffeine.mivote
 
 import android.os.Bundle
+import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.viewModels
-import androidx.appcompat.app.AppCompatActivity
-import androidx.compose.material.Surface
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.lifecycle.lifecycleScope
 import com.fueledbycaffeine.mivote.data.VoterInfo
 import com.fueledbycaffeine.mivote.ui.theme.MIVoteTheme
-import com.fueledbycaffeine.mivote.ui.voter.VoterRegistrationStatusScreen
 import com.fueledbycaffeine.mivote.ui.voter.VoterRegistrationViewModel
+import com.fueledbycaffeine.mivote.ui.voter.status.VoterRegistrationStatusScreen
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
 import java.time.LocalDate
 
 @AndroidEntryPoint
-class MainActivity : AppCompatActivity() {
+class MainActivity : ComponentActivity() {
 
   // TODO: DELETE WHEN CAN ENTER
   private val voterInfo = VoterInfo(
@@ -39,12 +38,10 @@ class MainActivity : AppCompatActivity() {
 
     setContent {
       MIVoteTheme {
-        Surface {
-          VoterRegistrationStatusScreen(
-            voterInfo = voterInfo,
-            registrationViewModel = voterRegistrationViewModel
-          )
-        }
+        VoterRegistrationStatusScreen(
+          voterInfo = voterInfo,
+          registrationLiveData = voterRegistrationViewModel.registrationLiveData
+        )
       }
     }
   }
