+++ app/src/main/java/com/fueledbycaffeine/mivote/ui/core/LoadingIndicator.kt
@@ -1,0 +1,33 @@
+package com.fueledbycaffeine.mivote.ui.core
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.CircularProgressIndicator
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.dimensionResource
+import androidx.compose.ui.tooling.preview.Preview
+import com.fueledbycaffeine.mivote.R
+import com.fueledbycaffeine.mivote.ui.theme.MIVoteTheme
+
+@Composable
+fun LoadingIndicator() {
+  Column(
+    modifier = Modifier
+      .fillMaxSize()
+      .padding(dimensionResource(id = R.dimen.loading_indicator_top_margin)),
+    horizontalAlignment = Alignment.CenterHorizontally
+  ) {
+    CircularProgressIndicator()
+  }
+}
+
+@Preview
+@Composable
+fun PreviewLoadingIndicator() {
+  MIVoteTheme {
+    LoadingIndicator()
+  }
+}
