+++ app/src/main/res/values/strings.xml
@@ -8,4 +8,5 @@
     <string name="try_again">Try Again</string>
     <string name="already_registered">Awesome! You are registered to vote!</string>
     <string name="registration_continue">Continue</string>
+
 </resources>
